restructure photcal
